+++ include/libcopp/utils/config/stl_include_prefix.h
@@ -15,8 +15,24 @@
 #ifndef LIBCOPP_UTILS_CONFIG_STL_INCLUDE_PREFIX_H
 #define LIBCOPP_UTILS_CONFIG_STL_INCLUDE_PREFIX_H
 
-#if defined(__GNUC__) && !defined(_WIN32) && !defined(__CYGWIN__)
-#  pragma GCC visibility push(default)
+// ============= Patch for some Compilers's mistake =============
+#if defined(__GNUC__) && !defined(__clang__) && !defined(__apple_build_version__)
+#  if (__GNUC__ * 100 + __GNUC_MINOR__ * 10) >= 460
+#    pragma GCC diagnostic push
+#  endif
+
+#  if (__GNUC__ * 100 + __GNUC_MINOR__) == 1402
+#    pragma GCC diagnostic ignored "-Wuninitialized"
+#    pragma GCC diagnostic ignored "-Wmaybe-uninitialized"
+#  endif
+
+#  if defined(__SANITIZE_THREAD__)
+#    pragma GCC diagnostic ignored "-Wtsan"
+#  elif defined(__has_feature)
+#    if __has_feature(thread_sanitizer)
+#      pragma GCC diagnostic ignored "-Wtsan"
+#    endif
+#  endif
 #endif
 
 #endif