+++ include/libcopp/utils/config/stl_include_suffix.h
@@ -14,8 +14,10 @@
 
 #ifdef LIBCOPP_UTILS_CONFIG_STL_INCLUDE_PREFIX_H
 
-#  if defined(__GNUC__) && !defined(_WIN32) && !defined(__CYGWIN__)
-#    pragma GCC visibility pop
+#  if defined(__GNUC__) && !defined(__clang__) && !defined(__apple_build_version__)
+#    if (__GNUC__ * 100 + __GNUC_MINOR__ * 10) >= 460
+#      pragma GCC diagnostic pop
+#    endif
 #  endif
 
 #  undef LIBCOPP_UTILS_CONFIG_STL_INCLUDE_PREFIX_H