+++ test/cli/shell_font.cpp
@@ -29,7 +29,7 @@ namespace cli {
 namespace detail {
 static char tolower(char c) {
   if (c >= 'A' && c <= 'Z') {
-    return c - 'A' + 'a';
+    return static_cast<char>(c - 'A' + 'a');
   }
 
   return c;
@@ -106,7 +106,7 @@ std::string shell_font::GetStyleCode(int iFlag) {
     std::string base = "30";
     int iStart = 0;
     for (; iStart < 8 && !(iFlag & (1 << iStart)); ++iStart);
-    if (iStart < 8) base[1] += static_cast<char>(iStart);
+    if (iStart < 8) base[1] = static_cast<char>(base[1] + iStart);
     ret += std::string((!bFirst) ? ";" : "") + base;
     bFirst = false;
   }
@@ -117,7 +117,7 @@ std::string shell_font::GetStyleCode(int iFlag) {
     std::string base = "40";
     int iStart = 0;
     for (; iStart < 8 && !(iFlag & (1 << iStart)); ++iStart);
-    if (iStart < 8) base[1] += static_cast<char>(iStart);
+    if (iStart < 8) base[1] = static_cast<char>(base[1] + iStart);
     ret += std::string((!bFirst) ? ";" : "") + base;
     // bFirst = false; no need to set because not used later
   }